refactor: WebSocket Redis 수정
