+++ src/main/java/com/example/chatserver/handler/WebSocketChatHandler.java
@@ -4,8 +4,10 @@ import com.example.chatserver.dto.ChatDto;
 import com.example.chatserver.service.ChatRoomManager;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketSession;
@@ -18,6 +20,7 @@ public class WebSocketChatHandler extends TextWebSocketHandler {
 
     private final ChatRoomManager chatRoomManager;
     private final ObjectMapper objectMapper = new ObjectMapper();
+    private final StringRedisTemplate stringRedisTemplate;
 
     @Override
     protected void handleTextMessage(WebSocketSession session, TextMessage message) throws JsonProcessingException {
@@ -35,6 +38,19 @@ public class WebSocketChatHandler extends TextWebSocketHandler {
 
     private void enterChatRoom(ChatDto chatDto, WebSocketSession session) {
         log.info("enter chatDto : " + chatDto.toString());
-        chatRoomManager.enter(chatDto, session);
+        String channel = "chatRoom:"+chatDto.getChatRoomId();
+        subscribe(channel,session);
+        chatRoomManager.enter(chatDto, channel);
+
+    }
+
+    private void subscribe(String channel, WebSocketSession session) {
+        Objects.requireNonNull(stringRedisTemplate.getConnectionFactory())
+            .getConnection()
+            .subscribe(getMessageHandler(session), channel.getBytes());
+    }
+
+    private RedisMessageHandler getMessageHandler(WebSocketSession session) {
+        return new RedisMessageHandler(session);
     }
 }
