+++ src/test/java/com/example/chatserver/service/ChatRoomManagerTest.java
@@ -11,20 +11,15 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.mockito.MockitoAnnotations;
 import org.springframework.web.socket.WebSocketSession;
 
-@ExtendWith(MockitoExtension.class)
 public class ChatRoomManagerTest {
 
     @Mock
     private ChatRoomService chatRoomService;
-
-    @Mock
-    private WebSocketSession session;
 
     @InjectMocks
     private ChatRoomManager chatRoomManager;
@@ -33,6 +28,8 @@ public class ChatRoomManagerTest {
 
     @BeforeEach
     public void setUp() {
+        MockitoAnnotations.openMocks(this);
+
         chatDto = new ChatDto();
         chatDto.setUsername("testUser");
         chatDto.setChatRoomId(123L);
@@ -41,34 +38,29 @@ public class ChatRoomManagerTest {
 
     @Test
     public void testEnter(){
-        // Arrange
+        // Given
         String expectedChannel = "chatRoom:123";
         String expectedMessage = "testUser님이 입장하셨습니다.";
+        String channel = "1";
 
-        // Mockito의 any 메서드를 사용하여 임의의 WebSocketSession을 받아들이도록 설정
-        doNothing().when(chatRoomService).subscribe(eq(expectedChannel), any(WebSocketSession.class));
-        doNothing().when(chatRoomService).publish(eq(expectedChannel), any(String.class));
+        // When
+        chatRoomManager.enter(chatDto, channel);
 
-        // Act
-        chatRoomManager.enter(chatDto, session);
-
-        // Assert
-        verify(chatRoomService, times(1)).subscribe(eq(expectedChannel), any(WebSocketSession.class));
+        // Then
         chatDto.setMessage(expectedMessage); // Adjust message after calling enter
         verify(chatRoomService, times(1)).publish(eq(expectedChannel), eq(getTextMessage(chatDto)));
     }
 
     @Test
     public void testSendMessage(){
-        // Arrange
+        // Given
         String expectedChannel = "chatRoom:123";
-
         doNothing().when(chatRoomService).publish(eq(expectedChannel), any(String.class));
 
-        // Act
+        // When
         chatRoomManager.sendMessage(chatDto);
 
-        // Assert
+        // Then
         verify(chatRoomService, times(1)).publish(eq(expectedChannel), eq(getTextMessage(chatDto)));
     }
