+++ src/test/java/com/example/chatserver/service/ChatRoomServiceTest.java
@@ -1,20 +1,17 @@
 package com.example.chatserver.service;
 
-import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.*;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-import java.util.Objects;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.data.redis.connection.MessageListener;
-import org.springframework.data.redis.connection.RedisConnection;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.StringRedisConnection;
 import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.web.socket.WebSocketSession;
 
 public class ChatRoomServiceTest {
 
@@ -46,22 +43,5 @@ public class ChatRoomServiceTest {
 
         // Then
         verify(stringRedisTemplate, times(1)).convertAndSend(eq(channel), eq(message));
-    }
-
-    @Test
-    public void testSubscribe() {
-        // Given
-        String channel = "testChannel";
-        WebSocketSession session = mock(WebSocketSession.class);
-        RedisConnection redisConnection = mock(StringRedisConnection.class);
-
-        // Redis 연결을 Mocking
-        when(Objects.requireNonNull(stringRedisTemplate.getConnectionFactory()).getConnection()).thenReturn(redisConnection);
-
-        // When
-        chatRoomService.subscribe(channel, session);
-
-        // Then
-        verify(redisConnection, times(1)).subscribe(any(MessageListener.class), any(byte[].class));
     }
 }
